requirements.txt
appdirs==1.4.4
asgiref==3.7.2
asyncio==3.4.3
attrs==20.3.0
autopep8==2.0.2
awscli==1.27.117
backports.shutil-get-terminal-size==1.0.0
backports.shutil_which==3.5.2
beautifulsoup4==4.12.2
botocore==1.29.117
braceexpand==0.1.7
Brotli==1.0.9
cachelib==0.10.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
check50==3.3.7
click==8.1.3
colorama==0.4.4
compare50==1.2.6
contourpy==1.0.7
cryptography==40.0.2
cs50==9.2.5
CS50-VSIX-Client @ file:///tmp/cs50.vsix/python-clients/cs50vsix-client
cssselect2==0.7.0
cycler==0.11.0
data-reader==0.0.8
Django==4.2.2
docutils==0.16
EditorConfig==0.12.3
Flask==2.2.3
Flask-Session==0.4.0
fonttools==4.39.3
frozendict==2.3.8
greenlet==2.0.2
help50==3.0.5
html5lib==1.1
icdiff==2.0.6
idna==3.4
ijson==3.2.1
iniconfig==2.0.0
intervaltree==3.1.0
itsdangerous==2.1.2
jellyfish==0.11.2
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
jsbeautifier==1.14.7
kiwisolver==1.4.4
lib50==3.0.6
lxml==4.9.2
MarkupSafe==2.1.2
matplotlib==3.7.1
mplfinance==0.12.9b7
multitasking==0.0.11
natsort==8.3.1
numpy==1.24.2
packaging==23.1
pandas==2.0.1
pexpect==4.8.0
Pillow==9.5.0
plotly==5.15.0
pluggy==1.0.0
ptyprocess==0.7.0
pyasn1==0.5.0
pycodestyle==2.10.0
pycparser==2.21
pydyf==0.6.0
Pygments==2.15.1
pyparsing==3.0.9
PyPDF2==3.0.1
pyphen==0.14.0
pytest==7.3.1
python-dateutil==2.8.2
python-magic==0.4.27
pytz==2023.3
PyYAML==5.4.1
render50==9.1.1
requests==2.28.2
rsa==4.7.2
s3cmd==2.3.0
s3transfer==0.6.0
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
sortedcontainers==2.4.0
soupsieve==2.4.1
SQLAlchemy==1.4.46
sqlparse==0.4.4
style50==2.7.5
submit50==3.1.1
tenacity==8.2.2
termcolor==1.1.0
threadpoolctl==3.1.0
tinycss2==1.2.1
tqdm==4.65.0
tzdata==2023.3
urllib3==1.26.15
weasyprint==58.1
webencodings==0.5.1
websocket-client==1.6.1
websockets==11.0.3
Werkzeug==2.2.3
yfinance==0.2.20
zopfli==0.2.2
